perf: add dashboard loading states, lazy-load MBA dialogs, lazy crests
Acts on the high-impact items from the website review (see
docs/website-improvement-suggestions.md):

- Add a shared RouteLoadingState skeleton and loading.tsx for the
  snapshot-driven dashboard routes (NFL, NBA, MLB, F1, Fantasy F1,
  Premier League, La Liga, World Cup, SpaceX, News Pulse, Investments,
  Golf, Polling Aggregator, GitHub Trending). Client-side navigation
  now shows an instant editorial skeleton instead of a blank page.

- Code-split the two interaction-only MBA tracker modals
  (EmailDigestDialog, ApplicationEditDialog) into their own modules,
  lazy-loaded via next/dynamic and mounted only while open, so their
  JS leaves the initial bundle. Shared form types/helpers/styles moved
  to application-form.ts. Main client drops ~440 lines.

- Add loading="lazy"/decoding="async" to football CrestAvatar (rendered
  many-per-page in standings/fixtures) and decoding="async" to the
  already-lazy F1/Fantasy-F1 avatars.

Verified: tsc clean on changed files, eslint clean, MBA + football
suites pass (29 tests).
