+++ src/app/fantasy-formula-1/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function FantasyFormula1Loading() {
+  return <RouteLoadingState surfaceName="the Fantasy Formula 1 optimizer" />;
+}
