+++ src/app/formula-1/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function Formula1Loading() {
+  return <RouteLoadingState surfaceName="the Formula 1 dashboard" />;
+}
