+++ src/app/github-trending-pulse/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function GithubTrendingLoading() {
+  return <RouteLoadingState surfaceName="the GitHub trending dashboard" />;
+}
