+++ src/app/golf/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function GolfLoading() {
+  return <RouteLoadingState surfaceName="the golf leaderboard" />;
+}
