+++ src/app/investments/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function InvestmentsLoading() {
+  return <RouteLoadingState surfaceName="the investments tracker" />;
+}
