+++ src/app/la-liga/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function LaLigaLoading() {
+  return <RouteLoadingState surfaceName="the La Liga dashboard" />;
+}
