+++ src/app/mlb/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function MlbLoading() {
+  return <RouteLoadingState surfaceName="the MLB dashboard" />;
+}
