+++ src/app/nba/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function NbaLoading() {
+  return <RouteLoadingState surfaceName="the NBA dashboard" />;
+}
