+++ src/app/news-pulse/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function NewsPulseLoading() {
+  return <RouteLoadingState surfaceName="the news pulse dashboard" />;
+}
