+++ src/app/nfl/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function NflLoading() {
+  return <RouteLoadingState surfaceName="the NFL dashboard" />;
+}
