+++ src/app/polling-aggregator/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function PollingAggregatorLoading() {
+  return <RouteLoadingState surfaceName="the polling aggregator" />;
+}
