+++ src/app/premier-league/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function PremierLeagueLoading() {
+  return <RouteLoadingState surfaceName="the Premier League dashboard" />;
+}
