+++ src/app/spacex-mission-control/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function SpacexLoading() {
+  return <RouteLoadingState surfaceName="SpaceX mission control" />;
+}
