+++ src/app/world-cup-2026/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoadingState } from "@/components/RouteLoadingState";
+
+export default function WorldCupLoading() {
+  return <RouteLoadingState surfaceName="the World Cup dashboard" />;
+}
